+++ changeEventProducer/kafka/prometheus/prometheus.yml
@@ -12,7 +12,7 @@ scrape_configs:
     metrics_path: /metrics
     params:
       match[]:
-        - kafka_integration_success_count_total
+        - 'kafka_success_count'
 
   - job_name: 'failure_count'
     static_configs:
@@ -20,12 +20,5 @@ scrape_configs:
     metrics_path: /metrics
     params:
       match[]:
-        - kafka_integration_failure_count_total
+        - 'kafka_failure_count'
 
-  - job_name: 'latency'
-    static_configs:
-      - targets: ['localhost:9102']
-    metrics_path: /metrics
-    params:
-      match[]:
-        - kafka_integration_latency_seconds
