+++ config/config.yaml
@@ -24,5 +24,5 @@ db:
   # Connection string for postgres database
   connectionstring: "postgres://postgres@localhost:5432/db"
 
-producer:
+kafkaProducer:
   bootstrapServer: "localhost:9092"
