--- kafkaIntegration/prometheus/mapping.yml
@@ -1,16 +1,0 @@
-mappings:
-  - match: "kafkaIntegration.SuccessCount.*"
-    name: "kafka_integration_success_count_total"
-    type: counter
-    labels:
-      application: "kafka_integration"
-  - match: "kafkaIntegration.FailureCount.*"
-    name: "kafka_integration_failure_count_total"
-    type: counter
-    labels:
-      application: "kafka_integration"
-  - match: "kafkaIntegration.Latency.*"
-    name: "kafka_integration_latency_seconds"
-    type: gauge
-    labels:
-      application: "kafka_integration"
